S3 to Redshift/requirement.txt
configparser==5.0.0
boto3==1.14.7
psycopg2-binary==2.8.5
botocore==1.17.7
